requirements.txt
matplotlib==3.8.2
pandas==2.2.0
python-dotenv==1.0.1
selenium==4.17.2
